parsing complete

+++ libft/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 /*#include<stdio.h>*/
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (s[len] != '\0')
